mobile: Change Envoy Mobile's xDS to use EnvoyGrpc instead of GoogleGrpc (#30846)
This allows us to not have to import in the google-grpc library into
Envoy Mobile, limiting the binary size by up to 3.5MB.

Next, we'll add a new flag guard, ENABLE_XDS, instead of relying on
ENABLE_GOOGLE_GRPC to guard the xDS APIs, and remove APIs from
EngineBuilder that are  EngineBuilder that are no longer needed with
EnvoyGrpc.

+++ mobile/library/swift/EngineBuilder.swift
@@ -900,7 +900,7 @@ private extension EngineBuilder {
 #if ENVOY_GOOGLE_GRPC
     if let xdsBuilder = self.xdsBuilder {
       var cxxXdsBuilder = Envoy.Platform.XdsBuilder(xdsBuilder.xdsServerAddress.toCXX(),
-                                                    Int32(xdsBuilder.xdsServerPort))
+                                                    xdsBuilder.xdsServerPort)
       for (header, value) in xdsBuilder.xdsGrpcInitialMetadata {
         cxxXdsBuilder.addInitialStreamHeader(header.toCXX(), value.toCXX())
       }
